Refactor: extracted raw vaidation parsing logic to separate class. Extracted hard-coded specifications to another class.
